requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-extensions==2.2.7
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
phonenumberslite==8.12.46
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2019.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.14.0
sqlparse==0.3.0
tzdata==2022.1
urllib3==1.26.9
Werkzeug==1.0.0
whitenoise==5.2.0
